requirements.txt
Django
requests
webcolors
pyimgur
colormath
Pillow
gunicorn
psycopg2
django_compressor
djangorestframework
markdown
django-filter
django-extensions
